requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
langchain-text-splitters==0.0.1
langchain-community==0.0.10
pypdf==3.17.4
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1